refactor: tidy filters and node helpers
Clarify filter flows, consolidate redaction handling, and simplify node-side control paths.\nAdjust docs and suppressions for readability without changing external behavior.

+++ src/main/java/network/crypta/node/updater/NodeUpdateManager.java
@@ -369,7 +369,7 @@ public class NodeUpdateManager {
                     .nextInt(
                         (int)
                             SECONDS.toMillis(
-                                (long) Math.min(Math.pow(2, i), (double) MINUTES.toSeconds(15)))));
+                                (long) Math.min(Math.pow(2, i), MINUTES.toSeconds(15)))));
       } catch (InterruptedException _) {
         Thread.currentThread().interrupt();
       }
